+++ linetools/isgm/tests/files/test_joebvp_repr.joebvp
@@ -1,5 +1,5 @@
 specfile|restwave|zsys|col|bval|vel|nflag|bflag|vflag|vlim1|vlim2|wobs1|wobs2|z_comp|trans|rely|comment
-test.fits|1215.67000|2.92939000|13.3000|15.0000|0.|2|2|2|-300.0000|300.0000|4772.06378|4781.62408|2.92939000|HI|none|
-test.fits|1025.72220|2.92939000|13.3000|15.0000|0.|2|2|2|-300.0000|300.0000|4026.43132|4034.49783|2.92939000|HI|none|
-test.fits|1215.67000|2.92939000|13.3000|15.0000|0.|3|3|3|-300.0000|300.0000|4772.06378|4781.62408|2.92939000|HI|none|
-test.fits|1025.72220|2.92939000|13.3000|15.0000|0.|3|3|3|-300.0000|300.0000|4026.43132|4034.49783|2.92939000|HI|none|
+test.fits|1215.67000|2.92939000|13.3000|15.0000|0.0000|2|2|2|-300.0000|300.0000|4772.06378|4781.62408|2.92939000|HI|none|
+test.fits|1025.72220|2.92939000|13.3000|15.0000|0.0000|2|2|2|-300.0000|300.0000|4026.43132|4034.49783|2.92939000|HI|none|
+test.fits|1215.67000|2.92939000|13.3000|15.0000|0.0000|3|3|3|-300.0000|300.0000|4772.06378|4781.62408|2.92939000|HI|none|
+test.fits|1025.72220|2.92939000|13.3000|15.0000|0.0000|3|3|3|-300.0000|300.0000|4026.43132|4034.49783|2.92939000|HI|none|
