+++ linetools/spectra/tests/test_lsf_init.py
@@ -1,8 +1,5 @@
 # Module to run tests on spectra.lsf
 
-from __future__ import print_function, absolute_import, \
-     division, unicode_literals
-
 import os
 import pytest
 from astropy import units as u
