+++ linetools/spectra/xspectrum1d.py
@@ -1129,7 +1129,7 @@ class XSpectrum1D(object):
             table.add_column(coclm)
 
         # Write
-        table.write(outfil, format=format)
+        table.write(outfil, format=format, overwrite=True)
 
     def write_to_fits(self, outfil, select=False, clobber=True, fill_val=0.):
         """ Write to a multi-extension FITS file.
@@ -1218,9 +1218,6 @@ class XSpectrum1D(object):
         #
         if self.meta is not None and len(self.meta) > 0:
             prihdu.header['METADATA'] = meta_to_disk(self.meta)
-
-        from IPython import embed
-        embed(header='1223 of write_to_fits')
 
         # Units, etc.
         prihdu.header['NSPEC'] = self.nspec
