+++ linetools/tests/test_utils.py
@@ -1,7 +1,3 @@
-# Tests of linetools.utils
-
-from __future__ import print_function, absolute_import, division, unicode_literals
-
 # TEST_UNICODE_LITERALS
 
 import pytest
