+++ old_setup.cfg
@@ -1,0 +1,27 @@
+[build_sphinx]
+source-dir = docs
+build-dir = docs/_build
+all_files = 1
+
+[upload_docs]
+upload-dir = docs/_build/html
+show-response = 1
+
+[tool:pytest]
+minversion = 2.2
+norecursedirs = build docs/_build
+doctest_plus = enabled
+
+[ah_bootstrap]
+auto_use = True
+
+[metadata]
+package_name = linetools
+description = A package for analysing absorption and emission lines in 1d spectra, especially galaxy and quasar spectra.
+#long_description = A package for analysing 1d spectra, especially galaxy and quasar spectra.
+author = The linetools developers
+author_email = xavier@ucolick.org, neilcrighton@gmail.com
+license = BSD
+edit_on_github = True
+github_project = linetools/linetools
+url = http://linetools.readthedocs.org/
