+++ setup.cfg
@@ -1,27 +1,101 @@
-[build_sphinx]
-source-dir = docs
-build-dir = docs/_build
-all_files = 1
-
-[upload_docs]
-upload-dir = docs/_build/html
-show-response = 1
-
-[tool:pytest]
-minversion = 2.2
-norecursedirs = build docs/_build
-doctest_plus = enabled
-
-[ah_bootstrap]
-auto_use = True
-
 [metadata]
-package_name = linetools
+name = linetools
 description = A package for analysing absorption and emission lines in 1d spectra, especially galaxy and quasar spectra.
-#long_description = A package for analysing 1d spectra, especially galaxy and quasar spectra.
+#long_description = file: README.md
+#long_description_content_type = text/markdown
 author = The linetools developers
 author_email = xavier@ucolick.org, neilcrighton@gmail.com
-license = BSD
-edit_on_github = True
+license = BSD-3
 github_project = linetools/linetools
 url = http://linetools.readthedocs.org/
+edit_on_github = False
+keywords = linetools
+classifiers =
+    Development Status :: 4 - Beta
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: BSD License
+    Natural Language :: English
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Topic :: Documentation :: Sphinx
+    Topic :: Scientific/Engineering :: Astronomy
+    Topic :: Software Development :: Libraries :: Python Modules
+    Topic :: Software Development :: User Interfaces
+
+[options]
+zip_safe = False
+use_2to3=False
+packages = find:
+python_requires = >=3.8
+setup_requires = setuptools_scm
+include_package_data = True
+install_requires =
+    numpy>=1.20
+    astropy>=5.2.1
+    scipy>=1.6
+    h5py>=3.7
+    matplotlib>=3.3
+    PyYAML>=5.1
+    IPython>=7.10.0
+    qtpy>=1.9
+
+[options.extras_require]
+test =
+    pytest>=6.0.0
+    pytest-astropy
+    tox
+    pytest-cov
+    coverage
+    codecov
+docs =
+    sphinx
+    sphinx-automodapi
+    sphinx_rtd_theme
+dev =
+    pytest>=6.0.0
+    pytest-astropy
+    tox
+    pytest-cov
+    coverage
+    codecov
+    sphinx
+    sphinx-automodapi
+    sphinx_rtd_theme
+
+[options.package_data]
+* = *.rst, *.txt, data/*, data/*/*, data/*/*/*, data/*/*/*/*, data/*/*/*/*/*, data/*/*/*/*/*/*
+
+[options.entry_points]
+console_scripts =
+#
+    lt_absline = linetools.scripts.lt_absline:main
+    lt_line = linetools.scripts.lt_line:main
+    lt_xabssys = linetools.scripts.lt_xabssys:main
+    lt_xspec = linetools.scripts.lt_xspec:main
+    lt_continuumfit = linetools.scripts.lt_continuumfit:main
+    lt_plot = linetools.scripts.lt_plot:main
+    lt_radec = linetools.scripts.lt_radec:main
+    lt_solabnd = linetools.scripts.lt_solabnd:main
+    lt_get_COS_LP = linetools.scripts.lt_get_COS_LP:main
+
+[coverage:report]
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+    # Don't complain about packages we have installed
+    except ImportError
+    # Don't complain if tests don't hit assertions
+    raise AssertionError
+    raise NotImplementedError
+    # Don't complain about script hooks
+    def main\(.*\):
+    # Ignore branches that don't pertain to this version of Python
+    pragma: py{ignore_python_version}
+    # Don't complain about IPython completion helper
+    def _ipython_key_completions_
+
+[tool:pytest]
+testpaths = "linetools/tests"
