+++ setup.py
@@ -1,149 +1,64 @@
 #!/usr/bin/env python
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-import glob
+# NOTE: The configuration for the package, including the name, version, and
+# other information are set in the setup.cfg file.
+
 import os
 import sys
 
-import ah_bootstrap
 from setuptools import setup
 
-#A dirty hack to get around some early import/configurations ambiguities
-if sys.version_info[0] >= 3:
-    import builtins
-else:
-    import __builtin__ as builtins
-builtins._ASTROPY_SETUP_ = True
+from extension_helpers import get_extensions
 
-from astropy_helpers.setup_helpers import (
-    register_commands, get_debug_option, get_package_info)
-from astropy_helpers.git_helpers import get_git_devstr
-from astropy_helpers.version_helpers import generate_version_py
+# First provide helpful messages if contributors try and run legacy commands
+# for tests or docs.
 
-# Get some values from the setup.cfg
+TEST_HELP = """
+Note: running tests via 'python setup.py test' is now deprecated. The recommended method
+is to run:
+
+    tox -e test-alldeps
+
+The Python version can also be specified, e.g.:
+
+    tox -e py38-test-alldeps
+
+You can list all available environments by doing:
+
+    tox -a
+
+If you don't already have tox installed, you can install it by doing:
+
+    pip install tox
+
+If you want to run all or part of the test suite within an existing environment,
+you can use pytest directly:
+
+    pip install -e .[dev]
+    pytest
+
+For more information, see:
+
+  http://docs.astropy.org/en/latest/development/testguide.html#running-tests
+"""
+
+if 'test' in sys.argv:
+    print(TEST_HELP)
+    sys.exit(1)
+
+VERSION_TEMPLATE = """
+# Note that we need to fall back to the hard-coded version if either
+# setuptools_scm can't be imported or setuptools_scm can't determine the
+# version, so we catch the generic 'Exception'.
 try:
-    from ConfigParser import ConfigParser
-except ImportError:
-    from configparser import ConfigParser
+    from setuptools_scm import get_version
+    version = get_version(root='..', relative_to=__file__)
+except Exception:
+    version = '{version}'
+""".lstrip()
 
-conf = ConfigParser()
-conf.read(['setup.cfg'])
-metadata = dict(conf.items('metadata'))
-
-PACKAGENAME = metadata.get('package_name', 'package')
-DESCRIPTION = metadata.get('description', 'Astropy affiliated package')
-AUTHOR = metadata.get('author', '')
-AUTHOR_EMAIL = metadata.get('author_email', '')
-LICENSE = metadata.get('license', 'unknown')
-URL = metadata.get('url', 'http://astropy.org')
-
-# Get the long description from the package's docstring
-__import__(PACKAGENAME)
-package = sys.modules[PACKAGENAME]
-LONG_DESCRIPTION = package.__doc__
-
-# Store the package name in a built-in variable so it's easy
-# to get from other parts of the setup infrastructure
-builtins._ASTROPY_PACKAGE_NAME_ = PACKAGENAME
-
-# VERSION should be PEP386 compatible (http://www.python.org/dev/peps/pep-0386)
-VERSION = '0.3.dev'
-
-# Indicates if this version is a release version
-RELEASE = 'dev' not in VERSION
-
-if not RELEASE:
-    git_devstr = get_git_devstr(False)
-    if (git_devstr == '0') or (git_devstr == ''):
-        import linetools.version
-        version_version = linetools.version.version
-        VERSION = version_version
-    else:
-        VERSION = VERSION + git_devstr
-
-# Populate the dict of setup command overrides; this should be done before
-# invoking any other functionality from distutils since it can potentially
-# modify distutils' behavior.
-cmdclassd = register_commands(PACKAGENAME, VERSION, RELEASE)
-
-# Adjust the compiler in case the default on this platform is to use a
-# broken one.
-#adjust_compiler(PACKAGENAME)
-
-# Freeze build information in version.py
-generate_version_py(PACKAGENAME, VERSION, RELEASE,
-                    get_debug_option(PACKAGENAME))
-
-# Treat everything in scripts except README.rst as a script to be installed
-scripts = [fname for fname in glob.glob(os.path.join('scripts', '*'))
-           if os.path.basename(fname) != 'README.rst']
-
-
-# Get configuration information from all of the various subpackages.
-# See the docstring for setup_helpers.update_package_files for more
-# details.
-package_info = get_package_info()
-
-# Add the project-global data
-package_info['package_data'].setdefault(PACKAGENAME, [])
-
-# Need a recursive glob to find all package data files if there are
-# subdirectories (copied from specutils)
-import fnmatch
-def recursive_glob(basedir, pattern):
-    matches = []
-    for root, dirnames, filenames in os.walk(basedir):
-        for filename in fnmatch.filter(filenames, pattern):
-            matches.append(os.path.join(root, filename))
-    return matches
-
-data_files = recursive_glob(os.path.join(PACKAGENAME, 'data'), '*')
-data_files = [f[len(PACKAGENAME)+1:] for f in data_files]
-package_info['package_data'][PACKAGENAME] += data_files
-
-# Define entry points for command-line scripts
-entry_points = {}
-entry_points['console_scripts'] = [
-    'lt_absline = linetools.scripts.lt_absline:main',
-    'lt_line = linetools.scripts.lt_line:main',
-    'lt_xabssys = linetools.scripts.lt_xabssys:main',
-    'lt_xspec = linetools.scripts.lt_xspec:main',
-    'lt_continuumfit = linetools.scripts.lt_continuumfit:main',
-    'lt_plot = linetools.scripts.lt_plot:main',
-    'lt_radec = linetools.scripts.lt_radec:main',
-    'lt_solabnd = linetools.scripts.lt_solabnd:main',
-    'lt_get_COS_LP = linetools.scripts.lt_get_COS_LP:main',
-
-#    'astropy-package-template-example = packagename.example_mod:main',
-]
-
-# Include all .c files, recursively, including those generated by
-# Cython, since we can not do this in MANIFEST.in with a "dynamic"
-# directory name.
-c_files = []
-for root, dirs, files in os.walk(PACKAGENAME):
-    for filename in files:
-        if filename.endswith('.c'):
-            c_files.append(
-                os.path.join(
-                    os.path.relpath(root, PACKAGENAME), filename))
-package_info['package_data'][PACKAGENAME].extend(c_files)
-
-setup(name=PACKAGENAME,
-      version=VERSION,
-      description=DESCRIPTION,
-      scripts=scripts,
-      requires=['astropy'],
-      install_requires=['astropy'],
-      provides=[PACKAGENAME],
-      author=AUTHOR,
-      author_email=AUTHOR_EMAIL,
-      license=LICENSE,
-      url=URL,
-      long_description=LONG_DESCRIPTION,
-      cmdclass=cmdclassd,
-      zip_safe=False,
-      use_2to3=False,
-      entry_points=entry_points,
-      **package_info
-)
+# Version was causing troubles
+setup(#use_scm_version={'write_to': os.path.join('zdm', 'version.py'),
+      #                 'write_to_template': VERSION_TEMPLATE},
+      ext_modules=get_extensions())
